rename account config entry inv to value; remove unused code
